git clone: add argument to specify object hash
Ideally the object format should be determined during cloning as
whatever the remote advertises. That would require us to use `git clone`
instead of first initializing an empty repo and then fetching.

Due to the way our fetching is performed, this isn't entirely trivial to
implement however. We support specifying an arbitrary set of branch/tag
patterns, and perform retries to filter out invalid refspecs. It is
unclear to me how to cleanly make use `git clone` (no "clone, read
format, delete, actual fetch").
